Messaging display
Messaging tab for prescribers is correctly displaying
