+++ app/src/main/java/com/example/demorxgo/ChatMessage.java
@@ -6,10 +6,22 @@ import java.util.*;
 
 import android.content.Intent;
 import android.os.Bundle;
+import android.text.TextUtils;
+import android.text.format.DateFormat;
 import android.view.View;
+import android.widget.Button;
+import android.widget.EditText;
+import android.widget.ListView;
+import android.widget.ProgressBar;
 import android.widget.TextView;
+import android.widget.Toast;
 import android.widget.Toolbar;
 
+import com.firebase.ui.database.FirebaseListAdapter;
+import com.google.android.gms.tasks.OnCompleteListener;
+import com.google.android.gms.tasks.Task;
+import com.google.android.material.floatingactionbutton.FloatingActionButton;
+import com.google.firebase.auth.AuthResult;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
 import com.google.firebase.database.DataSnapshot;
@@ -19,45 +31,74 @@ import com.google.firebase.database.FirebaseDatabase;
 import com.google.firebase.database.ValueEventListener;
 import com.google.firebase.firestore.auth.User;
 
-public class ChatMessage {
+public class ChatMessage extends AppCompatActivity {
 
-    private String messageText;
-    private String messageUser;
-    private long messageTime;
+    private FirebaseListAdapter<Messages> adapter;
 
-    public ChatMessage(String messageText, String messageUser) {
-        this.messageText = messageText;
-        this.messageUser = messageUser;
 
-        // Initialize to current time
-        messageTime = new Date().getTime();
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_messaging);
+
+        displayChatMessage();
+
+        FloatingActionButton fab =
+                (FloatingActionButton) findViewById(R.id.fab);
+
+        fab.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                EditText input = (EditText) findViewById(R.id.input);
+
+                // Read the input field and push a new instance
+                // of ChatMessage to the Firebase database
+                FirebaseDatabase.getInstance()
+                        .getReference()
+                        .push()
+                        .setValue(new Messages(input.getText().toString(),
+                                FirebaseAuth.getInstance()
+                                        .getCurrentUser()
+                                        .getDisplayName())
+                        );
+
+                // Clear the input
+                input.setText("");
+            }
+        });
     }
 
-    public ChatMessage(){
 
-    }
+        private void displayChatMessage() {
 
-    public String getMessageText() {
-        return messageText;
-    }
+            ListView listOfMessages = (ListView) findViewById(R.id.list_of_messages);
 
-    public void setMessageText(String messageText) {
-        this.messageText = messageText;
-    }
+            adapter = new FirebaseListAdapter<Messages>(this, Messages.class,
+                    R.layout.message, FirebaseDatabase.getInstance().getReference()) {
+                @Override
+                protected void populateView(View v, Messages model, int position) {
+                    // Get references to the views of message.xml
+                    TextView messageText = (TextView) v.findViewById(R.id.message_text);
+                    TextView messageUser = (TextView) v.findViewById(R.id.message_user);
+                    TextView messageTime = (TextView) v.findViewById(R.id.message_time);
 
-    public String getMessageUser() {
-        return messageUser;
-    }
+                    // Set their text
+                    messageText.setText(model.getMessageText());
+                    messageUser.setText(model.getMessageUser());
 
-    public void setMessageUser(String messageUser) {
-        this.messageUser = messageUser;
-    }
+                    // Format the date before showing it
+                    messageTime.setText(DateFormat.format("dd-MM-yyyy (HH:mm:ss)",
+                            model.getMessageTime()));
+                }
+            };
 
-    public long getMessageTime() {
-        return messageTime;
-    }
+            listOfMessages.setAdapter(adapter);
 
-    public void setMessageTime(long messageTime) {
-        this.messageTime = messageTime;
+        }
+
+    //back button function to go back to home screen
+    public void backToPrescriberHome(View view){
+        startActivity(new Intent(getApplicationContext(),PrescriberHome.class));
+        finish();
     }
-}
+    }
