+++ app/src/main/java/com/example/demorxgo/Prescriber.java
@@ -36,7 +36,7 @@ public class Prescriber extends AppCompatActivity {
     ProgressBar progressBar;
     FirebaseAuth fAuth;
 
-    private FirebaseListAdapter<ChatMessage> adapter;
+    //private FirebaseListAdapter<ChatMessage> adapter;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -50,29 +50,29 @@ public class Prescriber extends AppCompatActivity {
         fAuth = FirebaseAuth.getInstance();
         mLoginBtn = findViewById(R.id.login2);
 
-        FloatingActionButton fab =
-                (FloatingActionButton)findViewById(R.id.fab);
-
-        fab.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                EditText input = (EditText)findViewById(R.id.input);
-
-                // Read the input field and push a new instance
-                // of ChatMessage to the Firebase database
-               FirebaseDatabase.getInstance()
-                        .getReference()
-                        .push()
-                        .setValue(new ChatMessage(input.getText().toString(),
-                                FirebaseAuth.getInstance()
-                                        .getCurrentUser()
-                                        .getDisplayName())
-                        );
-
-                // Clear the input
-                input.setText("");
-            }
-        });
+        //FloatingActionButton fab =
+        //        (FloatingActionButton)findViewById(R.id.fab);
+//
+        //fab.setOnClickListener(new View.OnClickListener() {
+        //    @Override
+        //    public void onClick(View view) {
+        //        EditText input = (EditText)findViewById(R.id.input);
+//
+        //        // Read the input field and push a new instance
+        //        // of ChatMessage to the Firebase database
+        //       FirebaseDatabase.getInstance()
+        //                .getReference()
+        //                .push()
+        //                .setValue(new ChatMessage(input.getText().toString(),
+        //                        FirebaseAuth.getInstance()
+        //                                .getCurrentUser()
+        //                                .getDisplayName())
+        //                );
+//
+        //        // Clear the input
+        //        input.setText("");
+        //    }
+        //});
         //login button click listener
         mLoginBtn.setOnClickListener(new View.OnClickListener() {
             @Override
@@ -110,7 +110,7 @@ public class Prescriber extends AppCompatActivity {
                             Intent intent = new Intent(Prescriber.this, PrescriberHome.class);
                             startActivity(intent);
 
-                            displayChatMessage();
+                            //displayChatMessage();
 
                         } else {
                             //failed login
@@ -126,32 +126,32 @@ public class Prescriber extends AppCompatActivity {
 
     }
 
-    private void displayChatMessage() {
+   //private void displayChatMessage() {
 
-        ListView listOfMessages = (ListView)findViewById(R.id.list_of_messages);
+   //    ListView listOfMessages = (ListView)findViewById(R.id.list_of_messages);
 
-        adapter = new FirebaseListAdapter<ChatMessage>(this, ChatMessage.class,
-                R.layout.message, FirebaseDatabase.getInstance().getReference()) {
-            @Override
-            protected void populateView(View v, ChatMessage model, int position) {
-                // Get references to the views of message.xml
-                TextView messageText = (TextView)v.findViewById(R.id.message_text);
-                TextView messageUser = (TextView)v.findViewById(R.id.message_user);
-                TextView messageTime = (TextView)v.findViewById(R.id.message_time);
+   //    adapter = new FirebaseListAdapter<ChatMessage>(this, ChatMessage.class,
+   //            R.layout.message, FirebaseDatabase.getInstance().getReference()) {
+   //        @Override
+   //        protected void populateView(View v, ChatMessage model, int position) {
+   //            // Get references to the views of message.xml
+   //            TextView messageText = (TextView)v.findViewById(R.id.message_text);
+   //            TextView messageUser = (TextView)v.findViewById(R.id.message_user);
+   //            TextView messageTime = (TextView)v.findViewById(R.id.message_time);
 
-                // Set their text
-                messageText.setText(model.getMessageText());
-                messageUser.setText(model.getMessageUser());
+   //            // Set their text
+   //            messageText.setText(model.getMessageText());
+   //            messageUser.setText(model.getMessageUser());
 
-                // Format the date before showing it
-                messageTime.setText(DateFormat.format("dd-MM-yyyy (HH:mm:ss)",
-                        model.getMessageTime()));
-            }
-        };
+   //            // Format the date before showing it
+   //            messageTime.setText(DateFormat.format("dd-MM-yyyy (HH:mm:ss)",
+   //                    model.getMessageTime()));
+   //        }
+   //    };
 
-        listOfMessages.setAdapter(adapter);
+   //    listOfMessages.setAdapter(adapter);
 
-}
+    //}
 
     //functions to open register and back buttons
     public void prescriber_register(View view){
